First Working 2015 BEST version created by Alex
This is the 2014 code as modified by Alex Swearingen for the 2015 robot.
Note there is unused code from the 2014 robot still in this first
release.
He fixed the boom control DC motor and added servo controls for the
collection box release and for the sweeper arm on the  scooper.
Katherine and Mr. Dyson made edits to set angles on the sweeper arm.